% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{MCD}
\alias{MCD}
\title{Objet d analyse de deux variables quantitatives}
\usage{
MCD(x, y, var.nom = c("x", "y"))
}
\arguments{
\item{x}{Premiere variable de l analyse (quantitative)}

\item{y}{deuxieme variable de l analyse (quantitative)}

\item{var.nom}{Stocker ou Definir le nom des variables}
}
\description{
Objet d analyse de deux variables quantitatives
}
\examples{
MCD(x=c(153,165,176),y=c(57,65,73),var.nom=c("taille", "poids"))
}
