% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{plot.mcd}
\alias{plot.mcd}
\title{Fonction permettant d obtenir une representation graphique de l objet cree}
\usage{
\method{plot}{mcd}(objet)
}
\arguments{
\item{objet}{objet contenant les valeurs de la variable x et y sous forme de vecteur et cree  grâce a la fonction MCD}
}
\description{
Fonction permettant d obtenir une representation graphique de l objet cree
}
\examples{
objet<-MCD(x=c(3,5,6),y=c(7,5,3),var.nom=c("mois", "jour"))
fonction<-plot(objet)
}
