% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{prediction.mcd}
\alias{prediction.mcd}
\title{Prediction de Y en fonction de l objet et de X}
\usage{
\method{prediction}{mcd}(objet, x)
}
\arguments{
\item{objet}{objet contenant les valeurs de la variable x et y sous forme de vecteur et cree  grâce a la fonction MCD}

\item{x}{valeur de la variable explicative}
}
\description{
Prediction de Y en fonction de l objet et de X
}
