% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{prediction}
\alias{prediction}
\title{Title}
\usage{
prediction(objet, ...)
}
\arguments{
\item{objet}{objet contenant les valeurs de la variable x et y sous forme de vecteur et cree  grâce a la fonction MCD}

\item{...}{autres}
}
\description{
Title
}
