% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{print.mcd}
\alias{print.mcd}
\title{Affiche un resume de l objet}
\usage{
\method{print}{mcd}(objet)
}
\arguments{
\item{objet}{objet contenant les valeurs de la variable x et y sous forme de vecteur et cree  grace a la fonction MCD}
}
\description{
Affiche un resume de l objet
}
\examples{
objet<-MCD(x=c(3,5,6),y=c(7,5,3),var.nom=c("mois", "jour"))
fonction<-print(objet)
}
